requirements.txt
streamlit==1.28.1
openai==1.3.5
httpx==0.27.2
librosa==0.10.0
matplotlib==3.8.4
openpyxl==3.1.5
soundfile==0.12.1
numpy>=1.24,<2
python-dotenv==1.0.0
pyannote.audio==3.1.1
torch==2.2.2
torchaudio==2.2.2
huggingface_hub==0.23.2
